requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap4==2023.1
Django==4.2.10
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-jalali-date==1.1.0
django-js-asset==2.2.0
django-rosetta==0.10.0
idna==3.6
jdatetime==4.1.1
mysqlclient==2.2.3
pillow==10.2.0
polib==1.2.0
PyMySQL==1.1.0
requests==2.31.0
rosetta==0.3
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
